Bem.Miso.Utl.Utl

Description

the extra utilities for usage

Documentation

type Elem a = forall b. GenElem (View a) b

a view with a class of a full element

type Mix a = GenBlk (View a)

a view with a class of a full block and full element

type MkBlk a m = Reader m (GenBlk a)

a view with access to the model and a class of a full block and full element

type MkNoBlkModsBlk a m = Reader m (GenNoBlkModsBlk a)

a view with access to the model and a class of a block, without its modifiers, and full element

type MkNoElemModsBlk a m = Reader m (GenNoElemModsBlk a)

a view with access to the model and a class of a full block and element, without its modifiers

type MkNoModsBlk a m = Reader m (GenNoModsBlk a)

a view with access to the model and a class of a full block and element, without their modifiers

type MkRoot a m = Reader m (View a)

the root view

type NoBlkModsMix a = GenNoBlkModsBlk (View a)

a view with a class of a block, without its modifiers, and full element

type NoElemModsMix a = GenNoElemModsBlk (View a)

a view with a class of a full block and element, without its modifiers

type NoModsElem a = forall b. GenNoModsElem (View a) b

a view with a class of an element, without modifiers.

type NoModsMix a = GenNoModsBlk (View a)

a view with a class of a block and full element, without its modifiers