Bem.Miso.View.Mk.Mk

Description

Make views.

Documentation

blkElem :: HtmlElem a c -> c -> FromBlkElem (View a)

Make a view having a class of a block and element along with their modifiers.

elem :: HtmlElem a c -> c -> forall b. FromFullElem b (View a)

Make a view having a class of a element along with its modifiers.

mkBlkElem :: HtmlElem a c -> c -> Reader m (BlkElem a)

Make a view having access to the model and a class of a block and element along with their modifiers.

data BlkElem a

the wrapper through that the Bem.Utl.Intr.FromBlkElem type can be passed into a non-function type constructor