Bem.Miso.Utl.Utl

Description

high-level extra utilities

Documentation

newtype BlkElem a

the wrapper through that the Bem.Utl.Intr.FromBlkElem type can be passed into a non-function type constructor

Constructors

BlkElem (FromBlkElem (View a))

newtype NoModsBlkElem a

the wrapper through that the Bem.Utl.Intr.NoModsBlkElem type can be passed into the non-function type constructor

Constructors

NoModsBlkElem (FromNoModsBlkElem (View a))

newtype BlkNoModsElem a

the wrapper through that the Bem.Utl.Intr.FromBlkNoModsElem type can be passed into the non-function type constructor

Constructors

BlkNoModsElem (FromBlkNoModsElem (View a))

newtype NoModsBlkNoModsElem a

the wrapper through that the Bem.Utl.Intr.FromNoModsBlkNoModsElem type can be passed into the non-function type constructor

Constructors

NoModsBlkNoModsElem (FromNoModsBlkNoModsElem (View a))

blkNoModsElem :: HtmlElem a c -> c -> FromBlkNoModsElem (View a)

Make a view having a class of a block along with its modifiers and element without its modifiers.

mkBlkNoModsElem :: HtmlElem a c -> c -> Reader m (BlkNoModsElem a)

Make a view having access to the model and a class of a block along with its modifiers and element without its modifiers.

mkNoModsBlkElem :: HtmlElem a c -> c -> Reader m (NoModsBlkElem a)

Make a view having access to the model and a class of a block without its modifiers and element along with its modifiers.

mkNoModsBlkNoModsElem :: HtmlElem a c -> c -> Reader m (NoModsBlkNoModsElem a)

Make a view having access to the model and a class of a block and element without their modifiers.

noModsBlkElem :: HtmlElem a c -> c -> FromNoModsBlkElem (View a)

Make a view having a class of a block without its modifiers and element along with its modifiers.

noModsBlkNoModsElem :: HtmlElem a c -> c -> FromNoModsBlkNoModsElem (View a)

Make a view having a class of a block and element without their modifiers.

noModsElem :: HtmlElem a c -> c -> forall b. FromElem b (View a)

Make a view having a class of an element without its modifiers.